requirements.txt
livekit-agents[google]~=1.2
livekit-plugins-cartesia>=0.3.0
python-dotenv>=1.0.1
simpleaudio>=1.0.4
